products/ai_learning_platform/backend/requirements.txt
# AI学習プラットフォーム バックエンド依存関係

# FastAPI関連
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# データベース関連
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.12.1

# 認証関連
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 環境変数管理
python-dotenv==1.0.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# OpenAI API
openai==1.3.7

# HTTP関連
httpx==0.25.2
requests==2.31.0

# 開発・テスト関連
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
structlog
# その他
markdown==3.5.1  # Markdownレンダリング